transform tests

+++ __tests__/testMigration.ts
@@ -1,9 +1,9 @@
 import path from 'path';
 import tempy from 'tempy';
 import fs from 'fs-extra';
-import { createMigration, RegisterMigration } from '../';
 import globby from 'globby';
 import expect from 'expect';
+import { createMigration, RegisterMigration } from '../';
 
 export const testMigration = ({
   registerMigration,
@@ -12,9 +12,21 @@ export const testMigration = ({
   registerMigration: RegisterMigration;
   fixtures: string;
 }): void => {
+  if (!fs.existsSync(fixtures)) {
+    throw new Error(`fixture path ${fixtures} doesn't exist`);
+  }
+
   const workingDir = tempy.directory();
   const beforeDirectory = path.join(fixtures, 'before');
   const afterDirectory = path.join(fixtures, 'after');
+
+  if (!fs.existsSync(beforeDirectory)) {
+    throw new Error(`please create a "before" directory in ${fixtures}`);
+  }
+
+  if (!fs.existsSync(afterDirectory)) {
+    throw new Error(`please create a "after" directory in ${afterDirectory}`);
+  }
 
   fs.copySync(beforeDirectory, workingDir);
